[Templating] Fix CI
