+++ src/Symfony/Component/Templating/Tests/DelegatingEngineTest.php
@@ -14,7 +14,8 @@ namespace Symfony\Component\Templating\Tests;
 use PHPUnit\Framework\TestCase;
 use Symfony\Component\Templating\DelegatingEngine;
 use Symfony\Component\Templating\EngineInterface;
-use Symfony\Component\Templating\StreamingEngineInterface;
+use Symfony\Component\Templating\Tests\Fixtures\MyStreamingEngine;
+use Symfony\Component\Templating\Tests\Fixtures\TestEngine;
 
 /**
  * @group legacy
@@ -141,29 +142,5 @@ class DelegatingEngineTest extends TestCase
             ->willReturn($supports);
 
         return $engine;
-    }
-}
-
-interface MyStreamingEngine extends StreamingEngineInterface, EngineInterface
-{
-}
-
-class TestEngine implements EngineInterface
-{
-    public function render($name, array $parameters = []): string
-    {
-    }
-
-    public function exists($name): bool
-    {
-    }
-
-    public function supports($name): bool
-    {
-        return true;
-    }
-
-    public function stream()
-    {
     }
 }
