+++ src/Symfony/Component/Templating/Tests/Fixtures/MyStreamingEngine.php
@@ -1,0 +1,19 @@
+<?php
+
+/*
+ * This file is part of the Symfony package.
+ *
+ * (c) Fabien Potencier <fabien@symfony.com>
+ *
+ * For the full copyright and license information, please view the LICENSE
+ * file that was distributed with this source code.
+ */
+
+namespace Symfony\Component\Templating\Tests\Fixtures;
+
+use Symfony\Component\Templating\EngineInterface;
+use Symfony\Component\Templating\StreamingEngineInterface;
+
+interface MyStreamingEngine extends StreamingEngineInterface, EngineInterface
+{
+}
