+++ src/Symfony/Component/Templating/Tests/Fixtures/ProjectTemplateCacheLoader.php
@@ -1,0 +1,27 @@
+<?php
+
+/*
+ * This file is part of the Symfony package.
+ *
+ * (c) Fabien Potencier <fabien@symfony.com>
+ *
+ * For the full copyright and license information, please view the LICENSE
+ * file that was distributed with this source code.
+ */
+
+namespace Symfony\Component\Templating\Tests\Fixtures;
+
+use Symfony\Component\Templating\Loader\CacheLoader;
+
+class ProjectTemplateCacheLoader extends CacheLoader
+{
+    public function getDir()
+    {
+        return $this->dir;
+    }
+
+    public function getLoader()
+    {
+        return $this->loader;
+    }
+}
