+++ src/Symfony/Component/Templating/Tests/Fixtures/ProjectTemplateChainLoader.php
@@ -1,0 +1,22 @@
+<?php
+
+/*
+ * This file is part of the Symfony package.
+ *
+ * (c) Fabien Potencier <fabien@symfony.com>
+ *
+ * For the full copyright and license information, please view the LICENSE
+ * file that was distributed with this source code.
+ */
+
+namespace Symfony\Component\Templating\Tests\Fixtures;
+
+use Symfony\Component\Templating\Loader\ChainLoader;
+
+class ProjectTemplateChainLoader extends ChainLoader
+{
+    public function getLoaders()
+    {
+        return $this->loaders;
+    }
+}
