+++ src/Symfony/Component/Templating/Tests/Fixtures/ProjectTemplateHelper.php
@@ -1,0 +1,22 @@
+<?php
+
+/*
+ * This file is part of the Symfony package.
+ *
+ * (c) Fabien Potencier <fabien@symfony.com>
+ *
+ * For the full copyright and license information, please view the LICENSE
+ * file that was distributed with this source code.
+ */
+
+namespace Symfony\Component\Templating\Tests\Fixtures;
+
+use Symfony\Component\Templating\Helper\Helper;
+
+class ProjectTemplateHelper extends Helper
+{
+    public function getName(): string
+    {
+        return 'foo';
+    }
+}
