+++ src/Symfony/Component/Templating/Tests/Helper/HelperTest.php
@@ -12,7 +12,7 @@
 namespace Symfony\Component\Templating\Tests\Helper;
 
 use PHPUnit\Framework\TestCase;
-use Symfony\Component\Templating\Helper\Helper;
+use Symfony\Component\Templating\Tests\Fixtures\ProjectTemplateHelper;
 
 /**
  * @group legacy
@@ -24,13 +24,5 @@ class HelperTest extends TestCase
         $helper = new ProjectTemplateHelper();
         $helper->setCharset('ISO-8859-1');
         $this->assertSame('ISO-8859-1', $helper->getCharset(), '->setCharset() sets the charset set related to this helper');
-    }
-}
-
-class ProjectTemplateHelper extends Helper
-{
-    public function getName(): string
-    {
-        return 'foo';
     }
 }
