+++ src/Symfony/Component/Templating/Tests/Loader/ChainLoaderTest.php
@@ -12,9 +12,9 @@
 namespace Symfony\Component\Templating\Tests\Loader;
 
 use PHPUnit\Framework\TestCase;
-use Symfony\Component\Templating\Loader\ChainLoader;
 use Symfony\Component\Templating\Loader\FilesystemLoader;
 use Symfony\Component\Templating\TemplateReference;
+use Symfony\Component\Templating\Tests\Fixtures\ProjectTemplateChainLoader;
 
 /**
  * @group legacy
@@ -33,20 +33,20 @@ class ChainLoaderTest extends TestCase
 
     public function testConstructor()
     {
-        $loader = new ProjectTemplateLoader1([$this->loader1, $this->loader2]);
+        $loader = new ProjectTemplateChainLoader([$this->loader1, $this->loader2]);
         $this->assertEquals([$this->loader1, $this->loader2], $loader->getLoaders(), '__construct() takes an array of template loaders as its second argument');
     }
 
     public function testAddLoader()
     {
-        $loader = new ProjectTemplateLoader1([$this->loader1]);
+        $loader = new ProjectTemplateChainLoader([$this->loader1]);
         $loader->addLoader($this->loader2);
         $this->assertEquals([$this->loader1, $this->loader2], $loader->getLoaders(), '->addLoader() adds a template loader at the end of the loaders');
     }
 
     public function testLoad()
     {
-        $loader = new ProjectTemplateLoader1([$this->loader1, $this->loader2]);
+        $loader = new ProjectTemplateChainLoader([$this->loader1, $this->loader2]);
         $this->assertFalse($loader->load(new TemplateReference('bar', 'php')), '->load() returns false if the template is not found');
         $this->assertFalse($loader->load(new TemplateReference('foo', 'php')), '->load() returns false if the template does not exist for the given renderer');
         $this->assertInstanceOf(
@@ -54,13 +54,5 @@ class ChainLoaderTest extends TestCase
             $loader->load(new TemplateReference('foo.php', 'php')),
             '->load() returns a FileStorage if the template exists'
         );
-    }
-}
-
-class ProjectTemplateLoader1 extends ChainLoader
-{
-    public function getLoaders()
-    {
-        return $this->loaders;
     }
 }
