+++ src/Symfony/Component/Templating/Tests/Loader/FilesystemLoaderTest.php
@@ -13,9 +13,9 @@ namespace Symfony\Component\Templating\Tests\Loader;
 
 use PHPUnit\Framework\TestCase;
 use Psr\Log\LoggerInterface;
-use Symfony\Component\Templating\Loader\FilesystemLoader;
 use Symfony\Component\Templating\Storage\FileStorage;
 use Symfony\Component\Templating\TemplateReference;
+use Symfony\Component\Templating\Tests\Fixtures\ProjectTemplateFilesystemLoader;
 
 /**
  * @group legacy
@@ -32,27 +32,27 @@ class FilesystemLoaderTest extends TestCase
     public function testConstructor()
     {
         $pathPattern = self::$fixturesPath.'/templates/%name%.%engine%';
-        $loader = new ProjectTemplateLoader2($pathPattern);
+        $loader = new ProjectTemplateFilesystemLoader($pathPattern);
         $this->assertEquals([$pathPattern], $loader->getTemplatePathPatterns(), '__construct() takes a path as its second argument');
-        $loader = new ProjectTemplateLoader2([$pathPattern]);
+        $loader = new ProjectTemplateFilesystemLoader([$pathPattern]);
         $this->assertEquals([$pathPattern], $loader->getTemplatePathPatterns(), '__construct() takes an array of paths as its second argument');
     }
 
     public function testIsAbsolutePath()
     {
-        $this->assertTrue(ProjectTemplateLoader2::isAbsolutePath('/foo.xml'), '->isAbsolutePath() returns true if the path is an absolute path');
-        $this->assertTrue(ProjectTemplateLoader2::isAbsolutePath('c:\\\\foo.xml'), '->isAbsolutePath() returns true if the path is an absolute path');
-        $this->assertTrue(ProjectTemplateLoader2::isAbsolutePath('c:/foo.xml'), '->isAbsolutePath() returns true if the path is an absolute path');
-        $this->assertTrue(ProjectTemplateLoader2::isAbsolutePath('\\server\\foo.xml'), '->isAbsolutePath() returns true if the path is an absolute path');
-        $this->assertTrue(ProjectTemplateLoader2::isAbsolutePath('https://server/foo.xml'), '->isAbsolutePath() returns true if the path is an absolute path');
-        $this->assertTrue(ProjectTemplateLoader2::isAbsolutePath('phar://server/foo.xml'), '->isAbsolutePath() returns true if the path is an absolute path');
+        $this->assertTrue(ProjectTemplateFilesystemLoader::isAbsolutePath('/foo.xml'), '->isAbsolutePath() returns true if the path is an absolute path');
+        $this->assertTrue(ProjectTemplateFilesystemLoader::isAbsolutePath('c:\\\\foo.xml'), '->isAbsolutePath() returns true if the path is an absolute path');
+        $this->assertTrue(ProjectTemplateFilesystemLoader::isAbsolutePath('c:/foo.xml'), '->isAbsolutePath() returns true if the path is an absolute path');
+        $this->assertTrue(ProjectTemplateFilesystemLoader::isAbsolutePath('\\server\\foo.xml'), '->isAbsolutePath() returns true if the path is an absolute path');
+        $this->assertTrue(ProjectTemplateFilesystemLoader::isAbsolutePath('https://server/foo.xml'), '->isAbsolutePath() returns true if the path is an absolute path');
+        $this->assertTrue(ProjectTemplateFilesystemLoader::isAbsolutePath('phar://server/foo.xml'), '->isAbsolutePath() returns true if the path is an absolute path');
     }
 
     public function testLoad()
     {
         $pathPattern = self::$fixturesPath.'/templates/%name%';
         $path = self::$fixturesPath.'/templates';
-        $loader = new ProjectTemplateLoader2($pathPattern);
+        $loader = new ProjectTemplateFilesystemLoader($pathPattern);
         $storage = $loader->load(new TemplateReference($path.'/foo.php', 'php'));
         $this->assertInstanceOf(FileStorage::class, $storage, '->load() returns a FileStorage if you pass an absolute path');
         $this->assertEquals($path.'/foo.php', (string) $storage, '->load() returns a FileStorage pointing to the passed absolute path');
@@ -66,25 +66,12 @@ class FilesystemLoaderTest extends TestCase
         $logger = $this->createMock(LoggerInterface::class);
         $logger->expects($this->exactly(2))->method('debug');
 
-        $loader = new ProjectTemplateLoader2($pathPattern);
+        $loader = new ProjectTemplateFilesystemLoader($pathPattern);
         $loader->setLogger($logger);
         $this->assertFalse($loader->load(new TemplateReference('foo.xml', 'php')), '->load() returns false if the template does not exist for the given engine');
 
-        $loader = new ProjectTemplateLoader2([self::$fixturesPath.'/null/%name%', $pathPattern]);
+        $loader = new ProjectTemplateFilesystemLoader([self::$fixturesPath.'/null/%name%', $pathPattern]);
         $loader->setLogger($logger);
         $loader->load(new TemplateReference('foo.php', 'php'));
-    }
-}
-
-class ProjectTemplateLoader2 extends FilesystemLoader
-{
-    public function getTemplatePathPatterns()
-    {
-        return $this->templatePathPatterns;
-    }
-
-    public static function isAbsolutePath($path): bool
-    {
-        return parent::isAbsolutePath($path);
     }
 }
