+++ src/Symfony/Component/Templating/Tests/Storage/StorageTest.php
@@ -12,7 +12,7 @@
 namespace Symfony\Component\Templating\Tests\Storage;
 
 use PHPUnit\Framework\TestCase;
-use Symfony\Component\Templating\Storage\Storage;
+use Symfony\Component\Templating\Tests\Fixtures\TestStorage;
 
 /**
  * @group legacy
@@ -23,12 +23,5 @@ class StorageTest extends TestCase
     {
         $storage = new TestStorage('foo');
         $this->assertEquals('foo', (string) $storage, '__toString() returns the template name');
-    }
-}
-
-class TestStorage extends Storage
-{
-    public function getContent(): string
-    {
     }
 }
